IBN-2078 Fix circular dependency in osgi
Through decoupling the generic discovery from the specific manual discovery with
introduction of a new resource, the cyclic dependency in osgi is resolved.
