+++ bundles/io/org.eclipse.smarthome.io.rest.core/META-INF/MANIFEST.MF
@@ -17,7 +17,6 @@ Import-Package:
  com.google.common.base,
  com.google.common.collect,
  com.google.gson,
- com.google.gson.reflect;version="2.7.0",
  io.swagger.annotations;resolution:=optional,
  javax.annotation.security;resolution:=optional,
  javax.ws.rs,
@@ -25,7 +24,6 @@ Import-Package:
  javax.ws.rs.ext,
  org.apache.commons.lang,
  org.eclipse.jdt.annotation;resolution:=optional,
- org.eclipse.smarthome.binding.hue.handler,
  org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.config.core.dto,
  org.eclipse.smarthome.config.core.status,
