+++ bundles/io/org.eclipse.smarthome.io.rest.core/src/main/java/org/eclipse/smarthome/io/rest/core/internal/discovery/DiscoveryResource.java
@@ -18,22 +18,17 @@ import java.util.LinkedHashSet;
 import javax.annotation.security.RolesAllowed;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
-import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.UriInfo;
 
-import org.eclipse.smarthome.binding.hue.handler.HueBridgeDiscoveryHandler;
-import org.eclipse.smarthome.config.discovery.DiscoveryResult;
 import org.eclipse.smarthome.config.discovery.DiscoveryServiceRegistry;
 import org.eclipse.smarthome.config.discovery.ScanListener;
 import org.eclipse.smarthome.core.auth.Role;
-import org.eclipse.smarthome.io.rest.JSONResponse;
 import org.eclipse.smarthome.io.rest.RESTResource;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -71,8 +66,6 @@ public class DiscoveryResource implements RESTResource {
 
     private DiscoveryServiceRegistry discoveryServiceRegistry;
 
-    private HueBridgeDiscoveryHandler hueBridgeDiscoveryHandler;
-
     @Reference(cardinality = ReferenceCardinality.OPTIONAL, policy = ReferencePolicy.DYNAMIC)
     protected void setDiscoveryServiceRegistry(DiscoveryServiceRegistry discoveryServiceRegistry) {
         this.discoveryServiceRegistry = discoveryServiceRegistry;
@@ -80,15 +73,6 @@ public class DiscoveryResource implements RESTResource {
 
     protected void unsetDiscoveryServiceRegistry(DiscoveryServiceRegistry discoveryServiceRegistry) {
         this.discoveryServiceRegistry = null;
-    }
-
-    @Reference(cardinality = ReferenceCardinality.OPTIONAL, policy = ReferencePolicy.DYNAMIC)
-    protected void setHueBridgeDiscoveryHandler(HueBridgeDiscoveryHandler hueBridgeDiscoveryHandler) {
-        this.hueBridgeDiscoveryHandler = hueBridgeDiscoveryHandler;
-    }
-
-    protected void unsetHueBridgeDiscoveryHandler(HueBridgeDiscoveryHandler hueBridgeDiscoveryHandler) {
-        this.hueBridgeDiscoveryHandler = null;
     }
 
     @Context
@@ -126,23 +110,9 @@ public class DiscoveryResource implements RESTResource {
         return Response.ok(discoveryServiceRegistry.getMaxScanTimeout(bindingId)).build();
     }
 
-    @PUT
-    @Path("/{ip}")
-    @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "Adds a discovery.", response = String.class)
-    @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "OK", response = String.class) })
-    public Response putDiscovery(@PathParam("ip") @ApiParam(value = "ip") final String ip) {
-        DiscoveryResult bridge = this.hueBridgeDiscoveryHandler.addDiscovery(ip);
-        if (bridge == null) {
-            return JSONResponse.createErrorResponse(Status.NOT_FOUND, "Bridge not found");
-        }
-        return Response.ok(bridge).build();
-    }
-
     @Override
     public boolean isSatisfied() {
-        return discoveryServiceRegistry != null && hueBridgeDiscoveryHandler != null;
+        return discoveryServiceRegistry != null;
     }
 
 }
