+++ extensions/binding/org.eclipse.smarthome.binding.hue/META-INF/MANIFEST.MF
@@ -14,20 +14,25 @@ Import-Package:
  com.google.gson.annotations,
  com.google.gson.reflect,
  com.google.gson.stream,
+ io.swagger.annotations;resolution:=optional,
+ javax.annotation.security;resolution:=optional,
+ javax.ws.rs,
+ javax.ws.rs.core,
  org.apache.commons.io,
  org.apache.commons.lang,
  org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.smarthome.binding.hue,
- org.eclipse.smarthome.binding.hue.handler,
  org.eclipse.smarthome.config.core,
  org.eclipse.smarthome.config.core.status,
  org.eclipse.smarthome.config.discovery,
  org.eclipse.smarthome.config.discovery.upnp,
+ org.eclipse.smarthome.core.auth,
  org.eclipse.smarthome.core.library.types,
  org.eclipse.smarthome.core.thing,
  org.eclipse.smarthome.core.thing.binding,
  org.eclipse.smarthome.core.types,
  org.eclipse.smarthome.io.net.http,
+ org.eclipse.smarthome.io.rest,
  org.jupnp.model.meta,
  org.osgi.framework,
  org.slf4j
