--- extensions/binding/org.eclipse.smarthome.binding.hue/OSGI-INF/HueBridgeDiscoveryHandler.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2014,2017 Contributors to the Eclipse Foundation
-
-    See the NOTICE file(s) distributed with this work for additional
-    information regarding copyright ownership.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License 2.0 which is available at
-    http://www.eclipse.org/legal/epl-2.0
-
-    SPDX-License-Identifier: EPL-2.0
-
--->
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" immediate="true" name="org.eclipse.smarthome.binding.hue.handler.hue.discovery.handler">
-   <implementation class="org.eclipse.smarthome.binding.hue.handler.HueBridgeDiscoveryHandler"/>
-   <service>
-      <provide interface="org.eclipse.smarthome.binding.hue.handler.HueBridgeDiscoveryHandler"/>
-   </service>
-</scr:component>
