+++ extensions/binding/org.eclipse.smarthome.binding.hue/OSGI-INF/HueBridgeNupnpDiscovery.xml
@@ -17,6 +17,5 @@
    <implementation class="org.eclipse.smarthome.binding.hue.internal.discovery.HueBridgeNupnpDiscovery"/>
    <service>
       <provide interface="org.eclipse.smarthome.config.discovery.DiscoveryService"/>
-      <provide interface="org.eclipse.smarthome.binding.hue.internal.discovery.HueBridgeNupnpDiscovery"/>
    </service>
 </scr:component>
