+++ extensions/binding/org.eclipse.smarthome.binding.hue/OSGI-INF/ManualDiscoveryResource.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.smarthome.binding.hue.internal.discovery.ManualDiscoveryResource">
+   <service>
+      <provide interface="org.eclipse.smarthome.io.rest.RESTResource"/>
+      <provide interface="org.eclipse.smarthome.binding.hue.internal.discovery.ManualDiscoveryResource"/>
+   </service>
+   <implementation class="org.eclipse.smarthome.binding.hue.internal.discovery.ManualDiscoveryResource"/>
+</scr:component>
